OJ-3192 - Reorganise database utilities
