+++ lambdas/common/tests/retry.test.ts
@@ -1,4 +1,4 @@
-import { withRetry } from "../../src/util/retry";
+import { withRetry } from "../src/retry";
 
 jest.setTimeout(30000);
