+++ lambdas/database/src/exceptions/errors.ts
