+++ lambdas/database/src/util/is-record-expired.ts
